Flyttet over elementer i ducks til filstruktur med separate filer.
