+++ src/ducks/faktaavklaring/index.js
@@ -1,0 +1,15 @@
+import * as faktaavklaringActions from './actions';
+import * as faktaavklaringOperations from './operations';
+import * as faktaavklaringTypes from './types';
+import * as faktaavklaringSelectors from './selectors';
+
+import fagsakReducers from './reducers';
+
+export {
+  faktaavklaringActions,
+  faktaavklaringOperations,
+  faktaavklaringTypes,
+  faktaavklaringSelectors,
+};
+
+export default fagsakReducers;
