+++ src/ducks/faktaavklaring/types.js
@@ -1,0 +1,11 @@
+/**
+ * Types
+ * ----------------------------------------------------------------------------------
+ * Dette er action types som benyttes for å indikere hvordan fagsaker skal manipuleres
+ * eller oppdateres.
+ */
+
+export const OK = 'faktaavklaring/OK';
+export const FEILET = 'faktaavklaring/FEILET';
+export const PENDING = 'faktaavklaring/PENDING';
+export const OPPDATER_FAKTAAVKLARING = 'faktaavklaring/OPPDATER_FAKTAAVKLARING';
