+++ src/reducer.js
@@ -3,7 +3,7 @@ import { reducer as formReducer } from 'react-redux-form-validation';
 
 import fagsakerReducer from './ducks/fagsaker';
 import soknadReducer from './ducks/soknad';
-import faktaavklaringReducer from './ducks/faktaavklaring';
+import faktaavklaringReducer from './ducks/faktaavklaring/';
 import vurderingReducer from './ducks/vurdering';
 import nyesakerReducer from './ducks/nyesaker';
 import sakerbehandlesReducer from './ducks/sakerbehandles';
