+++ src/sider/saksbehandling.js
@@ -51,23 +51,10 @@ import {
 } from '../ducks/soknad';
 
 import {
-  hentFaktaavklaring,
-  sendFaktaavklaring,
-  oppdaterFaktaavklaringState,
-  FaktaavklaringSelector,
-  FaktaavklaringOppholdSelector,
-  FaktaavklaringOppholdPeriodeSelector,
-  FaktaavklaringSysselsettingSelector,
-  FaktaavklaringUtsendingSelector,
-  FaktaavklaringSektorSelector,
-  FaktaavklaringYrkesaktivitetFordelingSelector,
-  FaktaavklaringVirksomhetSelector,
-  FaktaavklaringAktivitetSelector,
-  FaktaavklaringBostedslandSelector,
-  FaktaavklaringTjenestemannSelector,
-  FaktaavklaringValgteArbeidsforholdSelector,
-  FaktaavklaringForretningsstedSelector,
-} from '../ducks/faktaavklaring';
+  faktaavklaringOperations,
+  faktaavklaringActions,
+  faktaavklaringSelectors,
+} from '../ducks/faktaavklaring/';
 
 import {
   hentVurdering,
@@ -247,7 +234,7 @@ const mapStateToProps = state => ({
   bekreftelser: BekreftelserSelector(state),
   oppsummering: OppsummeringSelector(state),
   soknad: SoknadSelector(state),
-  faktaavklaring: FaktaavklaringSelector(state),
+  faktaavklaring: faktaavklaringSelectors.FaktaavklaringSelector(state),
   soknadForm: SoknadenFormSelector(state),
   soknadArbeidsinntekt: ArbeidsinntektSelector(state),
   soknadOppholdUtland: OppholdUtlandSelector(state),
@@ -272,31 +259,31 @@ const mapStateToProps = state => ({
     kontaktEpost: ArbeidNorgeSelector(state).kontaktEpost,
     fullmektigFirma: ArbeidNorgeSelector(state).fullmektigFirma,
     fullmektigAdresse: ArbeidNorgeSelector(state).fullmektigAdresse,
-    faktaavklaringOppholdsLand: FaktaavklaringOppholdSelector(state).land,
-    faktaavklaringPeriodeFraOgMed: formatterDatoTilNorsk(FaktaavklaringOppholdPeriodeSelector(state).fom),
-    faktaavklaringPeriodeTilOgMed: formatterDatoTilNorsk(FaktaavklaringOppholdPeriodeSelector(state).tom),
-    faktaavklaringSysselsettingType: FaktaavklaringSysselsettingSelector(state).sysselsettingType,
-    faktaavklaringAnsattINorskSelskap: FaktaavklaringUtsendingSelector(state).ansattINorskSelskap,
-    faktaavklaringErstatterTidligereUtsendt: FaktaavklaringUtsendingSelector(state).erstatterTidligereUtsendt,
-    faktaavklaringUtsendingMindreEnn24Mnd: FaktaavklaringUtsendingSelector(state).utsendingMindreEnn24Mnd,
-    faktaavklaringForetakDriverINorge: FaktaavklaringUtsendingSelector(state).foretakDriverINorge,
-    faktaavklaringHarForutgaendeMedlemskap: FaktaavklaringUtsendingSelector(state).harForutgaendeMedlemskap,
-    faktaavklaringArbeidKnyttetTilVirksomhetUtlandet: FaktaavklaringUtsendingSelector(state).arbeidKnyttetTilVirksomhetUtlandet,
-    faktaavklaringSammeTypeVirksomhet: FaktaavklaringUtsendingSelector(state).sammeTypeVirksomhet,
-    faktaavklaringAnsattISektor: FaktaavklaringSektorSelector(state).ansattISektor,
-    faktaavklaringAntallLand: FaktaavklaringYrkesaktivitetFordelingSelector(state).antallLand,
-    faktaavklaringAktivitetINorge: FaktaavklaringVirksomhetSelector(state).aktivitetINorge,
-    faktaavklaringMarginaltArbeid: FaktaavklaringVirksomhetSelector(state).marginaltArbeid,
-    faktaavklaringVekslingMellomLand: FaktaavklaringVirksomhetSelector(state).vekslingMellomLand,
-    faktaavklaringAktivitetLand: FaktaavklaringAktivitetSelector(state).aktivitetLand,
-    faktaavklaringBekrefterFamiliebosted: FaktaavklaringBostedslandSelector(state).bekrefterFamiliebosted,
-    faktaavklaringBekrefterDisponering: FaktaavklaringBostedslandSelector(state).bekrefterFamiliebosted,
-    faktaavklaringBostedsland: FaktaavklaringBostedslandSelector(state).bostedsland,
-    faktaavklaringTjenestemann: FaktaavklaringTjenestemannSelector(state).tjenestemann,
-    faktaavklaringValgteArbeidsforhold: FaktaavklaringValgteArbeidsforholdSelector(state),
-    faktaavklaringForretningsstedLand: FaktaavklaringForretningsstedSelector(state).land,
-    faktaavklaringForretningsstedAntallArbeidsgivere: FaktaavklaringForretningsstedSelector(state).antallArbeidsgivere,
-    faktaavklaringForretningsstedFordelingArbeidsgivere: FaktaavklaringForretningsstedSelector(state).fordelingArbeidsgivere,
+    faktaavklaringOppholdsLand: faktaavklaringSelectors.FaktaavklaringOppholdSelector(state).land,
+    faktaavklaringPeriodeFraOgMed: formatterDatoTilNorsk(faktaavklaringSelectors.FaktaavklaringOppholdPeriodeSelector(state).fom),
+    faktaavklaringPeriodeTilOgMed: formatterDatoTilNorsk(faktaavklaringSelectors.FaktaavklaringOppholdPeriodeSelector(state).tom),
+    faktaavklaringSysselsettingType: faktaavklaringSelectors.FaktaavklaringSysselsettingSelector(state).sysselsettingType,
+    faktaavklaringAnsattINorskSelskap: faktaavklaringSelectors.FaktaavklaringUtsendingSelector(state).ansattINorskSelskap,
+    faktaavklaringErstatterTidligereUtsendt: faktaavklaringSelectors.FaktaavklaringUtsendingSelector(state).erstatterTidligereUtsendt,
+    faktaavklaringUtsendingMindreEnn24Mnd: faktaavklaringSelectors.FaktaavklaringUtsendingSelector(state).utsendingMindreEnn24Mnd,
+    faktaavklaringForetakDriverINorge: faktaavklaringSelectors.FaktaavklaringUtsendingSelector(state).foretakDriverINorge,
+    faktaavklaringHarForutgaendeMedlemskap: faktaavklaringSelectors.FaktaavklaringUtsendingSelector(state).harForutgaendeMedlemskap,
+    faktaavklaringArbeidKnyttetTilVirksomhetUtlandet: faktaavklaringSelectors.FaktaavklaringUtsendingSelector(state).arbeidKnyttetTilVirksomhetUtlandet,
+    faktaavklaringSammeTypeVirksomhet: faktaavklaringSelectors.FaktaavklaringUtsendingSelector(state).sammeTypeVirksomhet,
+    faktaavklaringAnsattISektor: faktaavklaringSelectors.FaktaavklaringSektorSelector(state).ansattISektor,
+    faktaavklaringAntallLand: faktaavklaringSelectors.FaktaavklaringYrkesaktivitetFordelingSelector(state).antallLand,
+    faktaavklaringAktivitetINorge: faktaavklaringSelectors.FaktaavklaringVirksomhetSelector(state).aktivitetINorge,
+    faktaavklaringMarginaltArbeid: faktaavklaringSelectors.FaktaavklaringVirksomhetSelector(state).marginaltArbeid,
+    faktaavklaringVekslingMellomLand: faktaavklaringSelectors.FaktaavklaringVirksomhetSelector(state).vekslingMellomLand,
+    faktaavklaringAktivitetLand: faktaavklaringSelectors.FaktaavklaringAktivitetSelector(state).aktivitetLand,
+    faktaavklaringBekrefterFamiliebosted: faktaavklaringSelectors.FaktaavklaringBostedslandSelector(state).bekrefterFamiliebosted,
+    faktaavklaringBekrefterDisponering: faktaavklaringSelectors.FaktaavklaringBostedslandSelector(state).bekrefterFamiliebosted,
+    faktaavklaringBostedsland: faktaavklaringSelectors.FaktaavklaringBostedslandSelector(state).bostedsland,
+    faktaavklaringTjenestemann: faktaavklaringSelectors.FaktaavklaringTjenestemannSelector(state).tjenestemann,
+    faktaavklaringValgteArbeidsforhold: faktaavklaringSelectors.FaktaavklaringValgteArbeidsforholdSelector(state),
+    faktaavklaringForretningsstedLand: faktaavklaringSelectors.FaktaavklaringForretningsstedSelector(state).land,
+    faktaavklaringForretningsstedAntallArbeidsgivere: faktaavklaringSelectors.FaktaavklaringForretningsstedSelector(state).antallArbeidsgivere,
+    faktaavklaringForretningsstedFordelingArbeidsgivere: faktaavklaringSelectors.FaktaavklaringForretningsstedSelector(state).fordelingArbeidsgivere,
   },
 });
 
@@ -305,11 +292,11 @@ const mapDispatchToProps = dispatch => ({
   hentFagsaker: saksnummer => dispatch(hentFagsaker(saksnummer)),
   hentSoknad: bid => dispatch(hentSoknad(bid)),
   sendSoknad: (bid, dokument) => dispatch(sendSoknad(bid, dokument)),
-  hentFaktaavklaring: saksnummer => dispatch(hentFaktaavklaring(saksnummer)),
-  sendFaktaavklaring: (bid, dokument) => dispatch(sendFaktaavklaring(bid, dokument)),
+  hentFaktaavklaring: saksnummer => dispatch(faktaavklaringOperations.hentFaktaavklaring(saksnummer)),
+  sendFaktaavklaring: (bid, dokument) => dispatch(faktaavklaringOperations.sendFaktaavklaring(bid, dokument)),
   hentVurdering: behandlingID => dispatch(hentVurdering(behandlingID)),
   oppdaterSoknad: values => { dispatch(oppdaterSoknadState(values)); },
-  oppdaterFaktaavklaring: values => { dispatch(oppdaterFaktaavklaringState(values)); },
+  oppdaterFaktaavklaring: values => { dispatch(faktaavklaringActions.oppdaterFaktaavklaringState(values)); },
 });
 
 const SaksbehandlingForm = validForm({
